ajout du scoring

+++ gpo3x/include/score.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-double getscore(graphnode** graph, uint16_t id,uint16_t depth );
+int getscore(graphnode** graph, uint16_t id,uint16_t depth);
